requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.2
africastalking==1.2.5